+++ app/src/main/java/org/breezyweather/common/basic/models/options/basic/UnitUtils.kt
@@ -49,6 +49,16 @@ import org.breezyweather.common.extensions.isTraditionalChinese
 import org.breezyweather.domain.settings.SettingsManager
 import java.text.FieldPosition
 import java.util.Locale
+import kotlin.Array
+import kotlin.Boolean
+import kotlin.Double
+import kotlin.Int
+import kotlin.Number
+import kotlin.String
+import kotlin.UnsupportedOperationException
+import kotlin.apply
+import kotlin.charArrayOf
+import kotlin.let
 import kotlin.math.pow
 import kotlin.math.roundToInt
 import kotlin.time.Duration.Companion.days