+++ app/src/main/java/org/breezyweather/ui/main/adapters/main/holder/DailyViewHolder.kt
@@ -17,7 +17,6 @@
 package org.breezyweather.ui.main.adapters.main.holder
 
 import android.annotation.SuppressLint
-import android.os.Build
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -29,6 +28,7 @@ import com.google.android.material.button.MaterialButton
 import com.google.android.material.button.MaterialButtonGroup
 import org.breezyweather.R
 import org.breezyweather.common.basic.BreezyActivity
+import org.breezyweather.common.extensions.getThemeColor
 import org.breezyweather.common.extensions.isLandscape
 import org.breezyweather.domain.settings.SettingsManager
 import org.breezyweather.ui.common.adapters.ButtonAdapter
@@ -36,22 +36,16 @@ import org.breezyweather.ui.common.widgets.RecyclerViewNoVerticalScrollTouchList
 import org.breezyweather.ui.common.widgets.trend.TrendRecyclerView
 import org.breezyweather.ui.main.adapters.trend.DailyTrendAdapter
 import org.breezyweather.ui.main.layouts.TrendHorizontalLinearLayoutManager
-import org.breezyweather.ui.main.utils.MainThemeColorProvider
 import org.breezyweather.ui.main.widgets.TrendRecyclerViewScrollBar
 import org.breezyweather.ui.theme.ThemeManager
 import org.breezyweather.ui.theme.resource.providers.ResourceProvider
-import org.breezyweather.ui.theme.weatherView.WeatherViewController
 
 class DailyViewHolder(parent: ViewGroup) : AbstractMainCardViewHolder(
     LayoutInflater.from(parent.context).inflate(R.layout.container_main_daily_trend_card, parent, false)
 ) {
-    private val title: TextView = itemView.findViewById(R.id.container_main_daily_trend_card_title)
-    private val subtitle: TextView = itemView.findViewById(R.id.container_main_daily_trend_card_subtitle)
-    private val buttonGroup: MaterialButtonGroup =
-        itemView.findViewById(R.id.container_main_daily_trend_card_buttonView)
-    private val trendRecyclerView: TrendRecyclerView = itemView.findViewById(
-        R.id.container_main_daily_trend_card_trendRecyclerView
-    )
+    private val subtitle: TextView = itemView.findViewById(R.id.daily_block_subtitle)
+    private val buttonGroup: MaterialButtonGroup = itemView.findViewById(R.id.daily_block_button_group)
+    private val trendRecyclerView: TrendRecyclerView = itemView.findViewById(R.id.daily_block_trendRecyclerView)
     private val scrollBar = TrendRecyclerViewScrollBar()
 
     init {
@@ -66,24 +60,12 @@ class DailyViewHolder(parent: ViewGroup) : AbstractMainCardViewHolder(
         provider: ResourceProvider,
         listAnimationEnabled: Boolean,
         itemAnimationEnabled: Boolean,
-        firstCard: Boolean,
+        selectedTab: String?,
+        setSelectedTab: (String?) -> Unit,
     ) {
-        super.onBindView(activity, location, provider, listAnimationEnabled, itemAnimationEnabled, firstCard)
+        super.onBindView(activity, location, provider, listAnimationEnabled, itemAnimationEnabled)
 
         val weather = location.weather ?: return
-        val colors = ThemeManager
-            .getInstance(context)
-            .weatherThemeDelegate
-            .getThemeColors(
-                context,
-                WeatherViewController.getWeatherKind(location),
-                WeatherViewController.isDaylight(location)
-            )
-
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
-            title.isAccessibilityHeading = true
-        }
-        title.setTextColor(colors[0])
 
         if (weather.current?.dailyForecast.isNullOrEmpty()) {
             subtitle.visibility = View.GONE
@@ -100,6 +82,15 @@ class DailyViewHolder(parent: ViewGroup) : AbstractMainCardViewHolder(
                 override val name = it.getDisplayName(activity)
             }
         }.toMutableList()
+        selectedTab?.let { tab ->
+            buttonList.indexOfFirst { it.name == tab }.let {
+                if (it >= 0) {
+                    trendAdapter.selectedIndex = it
+                } else {
+                    setSelectedTab(null) // Reset
+                }
+            }
+        }
 
         if (buttonList.size < 2) {
             buttonGroup.visibility = View.GONE
@@ -117,6 +108,11 @@ class DailyViewHolder(parent: ViewGroup) : AbstractMainCardViewHolder(
                         buttonGroup.removeView(it)
                     }
             }
+            buttonGroup.children
+                .filter { it is MaterialButton && it.tag == MaterialButtonGroup.OVERFLOW_BUTTON_TAG }
+                .forEach {
+                    it.contentDescription = context.getString(R.string.action_more)
+                }
             buttonList.forEachIndexed { index, button ->
                 buttonGroup.addView(
                     MaterialButton(
@@ -129,6 +125,7 @@ class DailyViewHolder(parent: ViewGroup) : AbstractMainCardViewHolder(
                         isChecked = index == trendAdapter.selectedIndex
                         setOnClickListener {
                             trendAdapter.selectedIndex = index
+                            setSelectedTab(button.name)
                             buttonGroup.children
                                 .filter { it is MaterialButton && it.tag != MaterialButtonGroup.OVERFLOW_BUTTON_TAG }
                                 .forEach { button ->
@@ -147,12 +144,12 @@ class DailyViewHolder(parent: ViewGroup) : AbstractMainCardViewHolder(
                 minHeight = context.resources.getDimensionPixelSize(R.dimen.daily_trend_item_height)
             )
         trendRecyclerView.setLineColor(
-            MainThemeColorProvider.getColor(location, com.google.android.material.R.attr.colorOutline)
+            context.getThemeColor(com.google.android.material.R.attr.colorOutline)
         )
         trendRecyclerView.setTextColor(
             ContextCompat.getColor(
                 context,
-                if (MainThemeColorProvider.isLightTheme(context, location)) {
+                if (ThemeManager.isLightTheme(context, location)) {
                     R.color.colorTextGrey
                 } else {
                     R.color.colorTextGrey2nd
@@ -167,6 +164,6 @@ class DailyViewHolder(parent: ViewGroup) : AbstractMainCardViewHolder(
         weather.todayIndex?.let { todayIndex ->
             trendRecyclerView.scrollToPosition(todayIndex)
         }
-        scrollBar.resetColor(location)
+        scrollBar.resetColor(activity)
     }
 }