+++ app/src/main/java/org/breezyweather/ui/settings/activities/CardDisplayManageActivity.kt
@@ -34,6 +34,7 @@ import org.breezyweather.common.basic.BreezyActivity
 import org.breezyweather.common.basic.models.options.appearance.CardDisplay
 import org.breezyweather.common.bus.EventBus
 import org.breezyweather.common.extensions.doOnApplyWindowInsets
+import org.breezyweather.common.extensions.getThemeColor
 import org.breezyweather.common.utils.ColorUtils
 import org.breezyweather.databinding.ActivityCardDisplayManageBinding
 import org.breezyweather.domain.settings.SettingsChangedMessage
@@ -43,7 +44,6 @@ import org.breezyweather.ui.common.decorations.GridMarginsDecoration
 import org.breezyweather.ui.common.decorations.ListDecoration
 import org.breezyweather.ui.common.widgets.slidingItem.SlidingItemTouchCallback
 import org.breezyweather.ui.settings.adapters.CardDisplayAdapter
-import org.breezyweather.ui.theme.ThemeManager
 
 class CardDisplayManageActivity : BreezyActivity() {
     private lateinit var mBinding: ActivityCardDisplayManageBinding
@@ -107,8 +107,8 @@ class CardDisplayManageActivity : BreezyActivity() {
         mBinding.toolbar.setBackgroundColor(
             ColorUtils.getWidgetSurfaceColor(
                 6f,
-                ThemeManager.getInstance(this).getThemeColor(this, androidx.appcompat.R.attr.colorPrimary),
-                ThemeManager.getInstance(this).getThemeColor(this, com.google.android.material.R.attr.colorSurface)
+                getThemeColor(androidx.appcompat.R.attr.colorPrimary),
+                getThemeColor(com.google.android.material.R.attr.colorSurface)
             )
         )
         mBinding.toolbar.setNavigationOnClickListener { finish() }
@@ -134,7 +134,7 @@ class CardDisplayManageActivity : BreezyActivity() {
         mBinding.recyclerView.addItemDecoration(
             ListDecoration(
                 this,
-                ThemeManager.getInstance(this).getThemeColor(this, com.google.android.material.R.attr.colorOutline)
+                getThemeColor(com.google.android.material.R.attr.colorOutline)
             )
         )
         mBinding.recyclerView.adapter = mCardDisplayAdapter
@@ -154,15 +154,6 @@ class CardDisplayManageActivity : BreezyActivity() {
         for (tag in otherTags) {
             tagList.add(CardTag(tag))
         }
-        val colors = ThemeManager.getInstance(this).getThemeColors(
-            this,
-            intArrayOf(
-                com.google.android.material.R.attr.colorOnPrimaryContainer,
-                com.google.android.material.R.attr.colorOnSecondaryContainer,
-                com.google.android.material.R.attr.colorPrimaryContainer,
-                com.google.android.material.R.attr.colorSecondaryContainer
-            )
-        )
         mTagAdapter = TagAdapter(
             tagList,
             { _: Boolean, _: Int, newPosition: Int ->