+++ app/src/main/java/org/breezyweather/ui/settings/activities/HourlyTrendDisplayManageActivity.kt
@@ -33,6 +33,7 @@ import org.breezyweather.R
 import org.breezyweather.common.basic.BreezyActivity
 import org.breezyweather.common.basic.models.options.appearance.HourlyTrendDisplay
 import org.breezyweather.common.extensions.doOnApplyWindowInsets
+import org.breezyweather.common.extensions.getThemeColor
 import org.breezyweather.common.utils.ColorUtils
 import org.breezyweather.databinding.ActivityHourlyTrendDisplayManageBinding
 import org.breezyweather.domain.settings.SettingsManager
@@ -41,7 +42,6 @@ import org.breezyweather.ui.common.decorations.GridMarginsDecoration
 import org.breezyweather.ui.common.decorations.ListDecoration
 import org.breezyweather.ui.common.widgets.slidingItem.SlidingItemTouchCallback
 import org.breezyweather.ui.settings.adapters.HourlyTrendDisplayAdapter
-import org.breezyweather.ui.theme.ThemeManager
 
 class HourlyTrendDisplayManageActivity : BreezyActivity() {
     private lateinit var mBinding: ActivityHourlyTrendDisplayManageBinding
@@ -105,8 +105,8 @@ class HourlyTrendDisplayManageActivity : BreezyActivity() {
         mBinding.toolbar.setBackgroundColor(
             ColorUtils.getWidgetSurfaceColor(
                 6f,
-                ThemeManager.getInstance(this).getThemeColor(this, androidx.appcompat.R.attr.colorPrimary),
-                ThemeManager.getInstance(this).getThemeColor(this, com.google.android.material.R.attr.colorSurface)
+                getThemeColor(androidx.appcompat.R.attr.colorPrimary),
+                getThemeColor(com.google.android.material.R.attr.colorSurface)
             )
         )
         mBinding.toolbar.setNavigationOnClickListener { finish() }
@@ -132,7 +132,7 @@ class HourlyTrendDisplayManageActivity : BreezyActivity() {
         mBinding.recyclerView.addItemDecoration(
             ListDecoration(
                 this,
-                ThemeManager.getInstance(this).getThemeColor(this, com.google.android.material.R.attr.colorOutline)
+                getThemeColor(com.google.android.material.R.attr.colorOutline)
             )
         )
         mBinding.recyclerView.adapter = mHourlyTrendDisplayAdapter
@@ -152,15 +152,6 @@ class HourlyTrendDisplayManageActivity : BreezyActivity() {
         for (tag in otherTags) {
             tagList.add(HourlyTrendTag(tag))
         }
-        val colors = ThemeManager.getInstance(this).getThemeColors(
-            this,
-            intArrayOf(
-                com.google.android.material.R.attr.colorOnPrimaryContainer,
-                com.google.android.material.R.attr.colorOnSecondaryContainer,
-                com.google.android.material.R.attr.colorPrimaryContainer,
-                com.google.android.material.R.attr.colorSecondaryContainer
-            )
-        )
         mTagAdapter = TagAdapter(
             tagList,
             { _: Boolean, _: Int, newPosition: Int ->